Extract Resonite material contract mapper
